Move Pair class to its own file and create decorator
