+++ fast_stream_22/matching/models.py
@@ -155,133 +155,3 @@ class Role(BaseClass):
 
     def from_anywhere(self) -> bool:
         return not {"Available Nationally", "Remote"}.isdisjoint(self.locations)
-
-
-class Pair:
-    scoring_weights: dict[str, int] = {
-        "first_location": 10,
-        "second_location": 5,
-        "department": 10,
-        "skill": 20,
-        "stretch": 10,
-    }
-
-    def __init__(self, c: Candidate, r: Role):
-        self.candidate = c
-        self.role = r
-        self._score: int = 0
-        self._disqualified = False
-
-    def score_pair(self) -> int:
-        self._check_location()
-        self._score_location()
-        self._score_clearance()
-        self._score_department()
-        self._ethical_check()
-        self._appropriate_for_year_group()
-        self._score_skill()
-        self._stretch_check()
-        self._check_nationality()
-        self._check_passport()
-        self._check_score()
-        return self._score
-
-    def _check_location(self):
-        if not self.candidate.can_relocate:
-            self._disqualified = not (
-                self.role.from_anywhere()
-                or self.candidate.first_preference_location in self.role.locations
-                or self.candidate.second_preference_location in self.role.locations
-            )
-
-    def _score_location(self):
-        if self.role.from_anywhere():
-            self._score += self.scoring_weights["first_location"]
-        elif (
-            self.candidate.first_preference_location in self.role.locations
-            or self.candidate.first_preference_location == "Any"
-        ):
-            self._score += self.scoring_weights["first_location"]
-        elif (
-            self.candidate.second_preference_location in self.role.locations
-            or self.candidate.second_preference_location == "Any"
-        ):
-            self._score += self.scoring_weights["second_location"]
-
-    def _score_clearance(self):
-        self.disqualified = self.candidate.clearance < self.role.clearance
-
-    def _check_nationality(self):
-        """
-        Nationality requirements come in three flavours: British national only, dual national, no restriction
-
-        :return:
-        """
-        self.disqualified = (
-            self.role.nationality_requirement > self.candidate.british_national
-        )
-
-    def _check_passport(self):
-        self.disqualified = (
-            self.role.passport_requirement and not self.candidate.has_passport
-        )
-
-    def _score_department(self):
-        if self.role.department not in self.candidate.prior_departments:
-            self._score += self.scoring_weights["department"]
-
-    def _ethical_check(self):
-        self.disqualified = (
-            self.candidate.no_immigration and self.role.immigration_role
-        ) or (self.candidate.no_defence and self.role.defence_role)
-
-    def _appropriate_for_year_group(self):
-        self.disqualified = (
-            self.candidate.year_group not in self.role.suitable_year_groups
-        )
-
-    def _score_skill(self) -> None:
-        if self.role.secondary_focus == self.candidate.last_role_secondary_skill:
-            self._score -= 5
-        bonus = self.scoring_weights["skill"]
-        for skill in (self.candidate.primary_skill, self.candidate.secondary_skill):
-            for focus in (self.role.skill_focus, self.role.secondary_focus):
-                if skill == focus:
-                    self._score += bonus
-                bonus -= 5
-
-    def _disqualify_skill(self) -> None:
-        if self.candidate.last_role_main_skill == self.role.skill_focus:
-            self.disqualified = True
-
-    def _stretch_check(self):
-        if (
-            not self.candidate.wants_private_office and self.role.private_office_role
-        ) or (
-            not self.candidate.wants_line_management and self.role.line_management_role
-        ):
-            self.disqualified = True
-        else:
-            if self.candidate.wants_private_office and self.role.private_office_role:
-                self._score += self.scoring_weights["stretch"]
-            if self.candidate.wants_line_management and self.role.line_management_role:
-                self._score += self.scoring_weights["stretch"]
-
-    def _check_score(self):
-        if self.score < 20:
-            self.disqualified = True
-
-    @property
-    def disqualified(self):
-        return self._disqualified
-
-    @disqualified.setter
-    def disqualified(self, value: bool):
-        if self._disqualified or value:
-            self._disqualified = True
-        else:
-            self._disqualified = False
-
-    @property
-    def score(self):
-        return self._score
